fix: bin tests
