+++ arbiter-core/src/tests/environment_control.rs
@@ -1,3 +1,5 @@
+use crate::environment::fork::ForkedDB;
+
 use super::*;
 
 #[tokio::test]
@@ -220,4 +222,25 @@ async fn stop_environment() {
     let (environment, client) = startup_user_controlled().unwrap();
     environment.stop().unwrap();
     assert!(deploy_arbx(client).await.is_err());
+}
+
+#[tokio::test]
+async fn fork_into_arbiter() {
+    let forked_db = ForkedDB::from_disk("./output/test.json").unwrap();
+
+    // Get the environment going
+    let environment = crate::environment::builder::EnvironmentBuilder::new()
+        .db(forked_db)
+        .build();
+
+    // Create a client
+    let client = RevmMiddleware::new(&environment, Some("name")).unwrap();
+
+    // Deal with the weth contract
+    let weth_address = Address::from_str("0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2").unwrap();
+    let weth = weth::WETH::new(weth_address, client.clone());
+
+    let address_to_check = Address::from_str("0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045").unwrap();
+    let balance = weth.balance_of(address_to_check).call().await.unwrap();
+    assert_eq!(balance, U256::from(34890707020710109111_u128));
 }
