+++ arbiter-core/src/tests/mod.rs
@@ -6,6 +6,7 @@ mod contracts;
 mod derives;
 mod environment_control;
 mod middleware_instructions;
+mod weth;
 
 use std::{str::FromStr, sync::Arc};
