--- bin/tests/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-pub mod weth;
